+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
 - Fix Ulatek Interrupt Alert counting up on the wrong cast
+- Add Interrupt Alert on Twin Fangs that alerts you when you receive a whisper